add AutoUpgrade option
On distributions that don't use versioned dependencies, it is extremely
easy for users to partially update their system to a broken state.  The
general wisdom to avoid breakage is to simply avoid refreshing the
database without also running a sysupgrade.  This makes -Qu completely
useless, prevents use of -Syuw for downloading updates, and can still
easily lead to problems if a sysupgrade is cancelled after the databases
are refreshed.  The more robust solution is to always run a sysupgrade
whenever a new package is installed.  This allows databases to be safely
refreshed, making -Qu and -Syuw functional again, and makes it much more
difficult to suffer a partial update.

Signed-off-by: Andrew Gregory <andrew.gregory.8@gmail.com>